Move form save on ctr+s to useForm hook
Since this is a common behaviour to all forms it is being moved to the
general hook to make it DRYer
